implementando novas validaçoes

+++ api/src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactRepository = require('../Repositories/ContactRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class ContactController {
   async index(req, res) {
@@ -10,6 +11,9 @@ class ContactController {
 
   async show(req, res) {
     const { id } = req.params;
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: 'Invalid id' });
+    }
     const contact = await ContactRepository.findById(id);
 
     if (!contact) {
@@ -22,22 +26,30 @@ class ContactController {
     const {
       name, email, phone, category_id,
     } = req.body;
+    if (email) {
+      const contactExists = await ContactRepository.findByEmail(email);
+      if (contactExists) {
+        return res.status(400).json({ error: 'o Email ja esta sendo usado' });
+      }
+    }
 
-    const contactExists = await ContactRepository.findByEmail(email);
+    if (category_id && !isValidUUID(category_id)) {
+      return res.status(400).json({ error: 'Invalid id in category' });
+    }
     if (!name) {
-      return res.sendStatus(400).json({ error: 'insira um nome' });
+      return res.status(400).json({ error: 'insira um nome' });
     }
     if (!phone) {
-      return res.sendStatus(400).json({ error: 'insira um telefone para contato' });
-    }
-    if (contactExists) {
-      return res.sendStatus(400).json({ error: 'o Email ja esta sendo usado' });
+      return res.status(400).json({ error: 'insira um telefone para contato' });
     }
 
     const contact = await ContactRepository.create({
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
-    res.send(contact);
+    return res.json(contact);
   }
 
   async update(req, res) {
@@ -45,20 +57,34 @@ class ContactController {
     const {
       name, email, phone, category_id,
     } = req.body;
-    const contactExists = await ContactRepository.findById(id);
-    if (!contactExists) {
-      return res.sendStatus(400).json({ error: 'esse id nao existe' });
+
+    if (category_id && !isValidUUID(category_id)) {
+      return res.status(400).json({ error: 'Invalid id in category' });
+    }
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: 'Invalid id' });
     }
     if (!name) {
-      return res.sendStatus(400).json({ error: 'insira um nome' });
+      return res.status(400).json({ error: 'insira um nome' });
     }
 
-    const contactWithEmail = await ContactRepository.findByEmail(email);
-    if (contactWithEmail && contactWithEmail.id !== id) {
-      return res.sendStatus(400).json({ error: 'o Email ja esta sendo usado' });
+    const contactExists = await ContactRepository.findById(id);
+
+    if (!contactExists) {
+      return res.status(400).json({ error: 'esse id nao existe' });
     }
+    if (email) {
+      const contactWithEmail = await ContactRepository.findByEmail(email);
+      if (contactWithEmail && contactWithEmail.id !== id) {
+        return res.status(400).json({ error: 'o Email ja esta sendo usado' });
+      }
+    }
+
     const contact = await ContactRepository.update(id, {
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
     res.json(contact);
   }
@@ -66,12 +92,14 @@ class ContactController {
   async delete(req, res) {
     const { id } = req.params;
     const contact = await ContactRepository.findById(id);
-
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: 'Invalid id' });
+    }
     if (!contact) {
       return res.status(404).json({ error: 'contato não encontrado!' });
     }
     await ContactRepository.delete(id);
-    res.sendStatus(204);
+    res.status(204);
   }
 }
